Update build to rollup
