+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": [ "es2015" ]
+  "presets": ["@babel/env"]
 }
