+++ package.json
@@ -3,10 +3,16 @@
   "description": "The Backbone Framework",
   "version": "4.0.0",
   "homepage": "https://marionettejs.com/",
-  "main": "lib/backbone.marionette.js",
-  "module": "lib/backbone.marionette.esm.js",
-  "jsnext:main": "lib/backbone.marionette.esm.js",
+  "browser": "dist/backbone.marionette.js",
+  "main": "dist/backbone.marionette.js",
+  "module": "dist/backbone.marionette.esm.js",
+  "jsnext:main": "dist/backbone.marionette.esm.js",
   "sideEffects": false,
+  "browserslist": [
+    ">0.5%",
+    "Explorer >= 10",
+    "not op_mini all"
+  ],
   "keywords": [
     "backbone",
     "plugin",
@@ -21,7 +27,7 @@
   ],
   "license": "MIT",
   "scripts": {
-    "build": "gulp build",
+    "build": "rollup -c --noConflict",
     "coverage": "gulp coverage",
     "coveralls": "gulp coveralls",
     "lint": "eslint --fix src/ && eslint --fix test/unit/",
@@ -51,11 +57,11 @@
     "underscore": ">= 1.8.3 <= 1.9.x"
   },
   "devDependencies": {
-    "babel-core": "6.25.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-register": "6.24.1",
     "backbone": "1.2.1 - 1.3.x",
     "chai": "3.5.0",
+    "@babel/core": "7.4.3",
+    "@babel/preset-env": "7.4.3",
+    "@babel/register": "7.4.0",
     "chai-jq": "0.0.9",
     "easy-sauce": "0.4.1",
     "gulp": "3.9.1",
@@ -74,18 +80,17 @@
     "lodash": "^4.17.3",
     "mocha": "3.1.2",
     "opn": "4.0.2",
-    "rollup": "0.45.2",
-    "rollup-plugin-babel": "2.7.1",
-    "rollup-plugin-commonjs": "5.0.5",
-    "rollup-plugin-json": "2.0.2",
-    "rollup-plugin-multi-entry": "2.0.1",
-    "rollup-plugin-node-globals": "1.0.9",
-    "rollup-plugin-node-resolve": "2.0.0",
     "run-sequence": "1.2.2",
     "sinon": "1.17.6",
     "sinon-chai": "2.8.0",
-    "uglify-js": "^3.0.27",
     "underscore": "1.8.3 - 1.9.x"
     "eslint": "5.16.0",
+    "rollup": "1.10.0",
+    "rollup-plugin-babel": "4.3.2",
+    "rollup-plugin-commonjs": "9.3.4",
+    "rollup-plugin-eslint": "5.1.0",
+    "rollup-plugin-json": "4.0.0",
+    "rollup-plugin-node-resolve": "4.2.3",
+    "rollup-plugin-terser": "4.0.4",
   }
 }
